+++ intg-appletv/driver.py
@@ -155,9 +155,14 @@ async def media_player_cmd_handler(
         return ucapi.StatusCodes.SERVICE_UNAVAILABLE
 
     # If the entity is OFF (device is in standby), we turn it on regardless of the actual command
+    # TODO #15 implement proper fix for correct entity OFF state (it may not remain in OFF state if connection is
+    #  established) + online check if we think it is in standby mode.
     if configured_entity.attributes[media_player.Attributes.STATE] == media_player.States.OFF:
         _LOG.debug("Device is off, sending turn on command")
-        return await device.turn_on()
+        # quick & dirty workaround for #15: the entity state is not always correct!
+        res = await device.turn_on()
+        if res != ucapi.StatusCodes.OK:
+            return res
 
     # Only proceed if device connection is established
     if device.is_on is False: